feat: updated messages list
in this part I updated the messages list

+++ app/src/main/java/pe/edu/upc/diligencetech/profiles/presentation/ProfileViewModel.kt
@@ -26,7 +26,7 @@ class ProfileViewModel @Inject constructor(
     private fun fetchUserData() {
         viewModelScope.launch {
             try {
-                _user.value = repository.getUserById(Constants.id!!.toInt()).data
+                _user.value = repository.getUserById(Constants.id!!).data
                 Log.d("ProfileViewModel", "User data fetched successfully: ${_user.value}")
             } catch (e: Exception) {
                 Log.e("ProfileViewModel", "Error fetching user data", e)
